🔨Removing mmap/huge-page allocations
The mmap() functions were leftovers from another project where I've
been directly allocating hugepages and have the system configured to
maintain a pool of permanent hugepages for the various systems that
run on it. This is not helpful to most folks, and I'm not doing any
zero-copy V4L2 stuff here (it's unrequired unlike with 1080p video)
so I've replaced those calls with simple mallocs here instead.